[ADD] fetch chat_id from users and alert from users preference report time
