+++ config/config.env
@@ -24,7 +24,7 @@ DATABASE_MASTER_DBNAME=crypto_watcher
 DATABASE_MASTER_SSLMODE=disable
 
 # scheduler configuration
-SCHEDULER_BITCOIN_FETCH=0 */5 * * * *
+SCHEDULER_CRYPTO_FETCH=0 */5 * * * *
 
 # coin api configuration
 COIN_API_HOST=https://rest.coinapi.io/v1
